+++ app/src/main/java/ru/practicum/android/diploma/ui/vacancy/VacancyViewModel.kt
@@ -4,8 +4,13 @@ import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
+import ru.practicum.android.diploma.data.dto.model.favorites.ShareData
+import ru.practicum.android.diploma.domain.api.VacancyInteractor
+import ru.practicum.android.diploma.domain.models.Vacancy
+import ru.practicum.android.diploma.util.SingleLiveEvent
 import ru.practicum.android.diploma.data.dto.model.VacancyFullItemDto
 import ru.practicum.android.diploma.domain.vacancy.VacancyInteractor
 
@@ -14,14 +19,24 @@ class VacancyViewModel(
     private val interactor: VacancyInteractor
 ) : AndroidViewModel(application) {
 
+    private val shareState = SingleLiveEvent<ShareData>()
+    fun observeShareState(): LiveData<ShareData> = shareState
     private val vacancyScreenStateLiveData = MutableLiveData<VacancyState>()
     private var listVacancy: VacancyFullItemDto? = null
 
+    private val _isFavorite = MutableLiveData<Boolean>()
+    val isFavorite: LiveData<Boolean> get() = _isFavorite
     val getVacancyScreenStateLiveData: LiveData<VacancyState> = vacancyScreenStateLiveData
 
+    fun shareVacancy(id: String) {
+        shareState.postValue(vacancyInteractor.getShareData(id))
+    }
+
+    fun isVacancyInFavorites(id: String) {
     fun getVacancyRessurces(id: String) {
         renderState(VacancyState.Loading)
         viewModelScope.launch {
+            _isFavorite.postValue(vacancyInteractor.isFavorite(id))
             interactor.getVacancyId(id).collect { pair ->
                 processResult(pair.first, pair.second)
             }
@@ -47,6 +62,14 @@ class VacancyViewModel(
                 renderState(VacancyState.Empty)
             }
 
+    fun onFavoriteClicked(vacancy: Vacancy) {
+        viewModelScope.launch {
+            if (_isFavorite.value == true) {
+                vacancyInteractor.deleteFavouritesVacancyEntity(vacancy)
+                _isFavorite.postValue(false)
+            } else {
+                vacancyInteractor.addVacancyToFavorites(vacancy)
+                _isFavorite.postValue(true)
             "Unknown Error" -> {
                 renderState(VacancyState.ServerError)
             }
